move FT selector data initialization into Helpers.cs class and remove old implementation from DataCacheHelper.cs class
